% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addorgname.R
\name{addorgname}
\alias{addorgname}
\title{addorgname}
\usage{
addorgname(
  data,
  CodeCol = "OrgId",
  NameCol = "OrgName",
  insert = TRUE,
  UseProxy = FALSE
)
}
\arguments{
\item{data}{data.frame containing the Organisation Codes to add Organisation Names to, unquoted string, no default}

\item{CodeCol}{the column name from data containing the organisation codes, quoted string, default "OrgId"}

\item{NameCol}{the name of the column to be added holding organisation names, quoted string, default "OrgName"}

\item{insert}{whether the new OrgName column should be positioned immediately after the CodeCol column (TRUE
or appended to the end of the data frame (FALSE), logical, default TRUE}

\item{UseProxy}{Whether to use proxy settings to connect to the directory.spineservices.nhs.uk/ODSAPISuite API.
If TRUE, the function will look for a UseProxy.csv file in the package library extdata directory.
The file must contain a single row of comma separated headers (url, port, username, password, auth)
and a second row specifying the values to be passed to each of these httr::use_proxy() function arguments.
All Argument headers must be present, leave values blank if not required. Quoted string, default FALSE}
}
\value{
returns the original data.frame with Organisation Name column added
}
\description{
Adds Organisation Names to a data.frame containing Organisation Codes using the NHS Digitial ODS ORD API.
}
\section{Notes}{
 View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{
# add OrgNm column to a data.frame containing a single column (OrgId) of Organisation Codes
codelist <- getODS(Status="Active",PostCode="S80B")[2]
addorgname(codelist)
addorgname(codelist, NameCol="MyNewColumnName", insert=FALSE)

}
\seealso{
Other odsR package functions: 
\code{\link{OrgRelLkpGP}()},
\code{\link{OrgRelLkpNHSTrust}()},
\code{\link{getODSfull}()},
\code{\link{getODS}()}
}
\concept{odsR package functions}
