% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{named_relationships}
\alias{named_relationships}
\title{Look-up table of named relationships in ODS}
\format{A dataframe of 6 rows and 6 variables
\describe{
  \item{id}{A relationship id}
  \item{meaning}{The nature of the relationship}
  \item{ex_id_from}{An example role id}
  \item{ex_role_from}{An example role name}
  \item{ex_id_to}{An example role id to which the relationship points}
  \item{ex_role_to}{An example role name to which the relationship points}
}}
\source{
HSCOrgRefData_DataItemCatalogue.pdf from \url{https://digital.nhs.uk/binaries/content/assets/website-assets/services/ods/xml-organisation-data-products/ancilliary.zip}
}
\usage{
named_relationships
}
\description{
A data set containing the meanings of the relationship ids and examples of
possible relationships. For more details see Appendix 1 from
HSCOrgRefData_DataItemCatalogue.pdf in
\url{https://digital.nhs.uk/binaries/content/assets/website-assets/services/ods/xml-organisation-data-products/ancilliary.zip}
}
\keyword{datasets}
