% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsR.R
\docType{package}
\name{odsR}
\alias{odsR}
\alias{odsR-package}
\title{odsR: A package for extracting NHS organisation data
from the NHS Digital ODS ORD API}
\description{
odsR: A package for extracting NHS organisation data
from the NHS Digital ODS ORD API
}
