% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrgParentLkp.R
\name{OrgParentLkp}
\alias{OrgParentLkp}
\title{OrgParentLkp}
\usage{
OrgParentLkp(PrimaryRole, NonPrimaryRole, ParentPrimaryRoles, FromDate)
}
\value{
returns a data.frame of  Organisation code, name, start date and end date plus the parent code and name
and associated start and end dates for the organisation >> parent relationship.
}
\description{
Generates an Organisation to Parent Lookup Data Frame from the NHS Digitial ODS ORD API.
}
\section{Notes}{
 View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{
# return GP Practice to CCG Lookup to include all organisations effective on or after 01-04-2013
OrgParentLkp("RO177","RO76","RO98","2013-04-01")

# return GP practice to CCG/PCT Lookup to include all organisations effective on or after 01-04-2012
OrgParentLkp("RO177","RO76",c("RO98","RO179"),"2012-04-01")

}
\seealso{
Other odsR package functions: \code{\link{AddOrgName}},
  \code{\link{getODSfull}}, \code{\link{getODS}}
}
