% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrgRelLkp.R
\name{OrgRelLkp}
\alias{OrgRelLkp}
\title{OrgRelLkp}
\usage{
OrgRelLkp(PrimaryRole, NonPrimaryRole, RelTypes, RelPrimaryRoles, FromDate)
}
\arguments{
\item{PrimaryRole}{The Primary Role code for organisations to be included in the lookup;
quoted string, no default}

\item{NonPrimaryRole}{The Non Primary Role code for organisations to be included in the lookup;
quoted string, no default}

\item{RelTypes}{The Named Relationship Types for related organisations to be included in the lookup;
character vector, no default}

\item{RelPrimaryRoles}{The Primary Role Ids for related organisations to be included in the lookup;
character vector, no default}

\item{FromDate}{The effective date from which to include Organisations operational on or after;
character string in the format "yyyy-mm-dd", no default}
}
\value{
returns a data.frame of Organisation codes, names, start dates and end dates plus the
organisations (code & name) they are related to and associated start and end dates
for the relationship.
}
\description{
Generates an organisational lookup table from the NHS Digital ODS ORD API
based on organisational relationship data.
}
\section{Notes}{
 This function is experimental and has only been thoroughly QAd for use with the arguments
         presented in the first example in order to generate GP Practice to CCG lookup tables. \cr \cr
         View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{
# return GP Practice to CCG Lookup to include all organisations effective on or after 01-04-2013

\dontrun{
OrgRelLkp(PrimaryRole    = "RO177",
          NonPrimaryRole = "RO76",
          RelTypes       = "RE4",
          RelPrimaryRoles = "RO98",
          FromDate       = "2013-04-01")
}

# return GP practice to CCG/PCT Lookup to include all organisations effective on or after 01-04-2012
\dontrun{
OrgRelLkp("RO177","RO76","RE4",c("RO98","RO179"),"2012-04-01")
}

}
\seealso{
Other odsR package functions: \code{\link{addorgname}},
  \code{\link{getODSfull}}, \code{\link{getODS}}
}
\concept{odsR package functions}
