% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrgRelLkp_GPPrac.R
\name{OrgRelLkpGP}
\alias{OrgRelLkpGP}
\title{OrgRelLkp}
\usage{
OrgRelLkpGP(
  PrimaryRole,
  NonPrimaryRole = "All",
  RelTypes = "None",
  RelPrimaryRoles = "None",
  FromDate,
  UseProxy = FALSE
)
}
\arguments{
\item{PrimaryRole}{The Primary Role code for organisations to be included in the lookup;
quoted string, no default}

\item{NonPrimaryRole}{The Non Primary Role code for organisations to be included in the lookup,
or 'All' to include all Non Primary Role Codes;
quoted string, default = "All"}

\item{RelTypes}{The Named Relationship Types for related organisations to be included in the lookup,
or 'None' to exclude Relationships from output;
character vector, default = "None"}

\item{RelPrimaryRoles}{The Primary Role Ids for related organisations to be included in the lookup,
or 'None' to exclude relationships from output;
character vector, default = "None"}

\item{FromDate}{The effective date from which to include Organisations operational on or after;
character string in the format "yyyy-mm-dd", no default}

\item{UseProxy}{Whether to use proxy settings to connect to the directory.spineservices.nhs.uk/ODSAPISuite API.
If TRUE, the function will look for a UseProxy.csv file in the package library extdata directory.
The file must contain a single row of comma separated headers (url, port, username, password, auth)
and a second row specifying the values to be passed to each of these httr::use_proxy() function arguments.
All Argument headers must be present, leave values blank if not required. Quoted string, default FALSE}
}
\value{
returns a data.frame of Organisation codes, names, start dates and end dates plus the start and end dates
that the organisation was in the specified role and related organisation codes & names and associated start and end dates
for the relationship.
}
\description{
Generates an organisational lookup table from the NHS Digital ODS ORD API
based on organisational relationship data.
}
\section{Notes}{
 This function is experimental and has only been thoroughly QAd for use with the arguments
         presented in the first example in order to generate GP Practice to CCG lookup tables. \cr \cr
         View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{

# return basic NHS Trust lookup without relationships to other organisation types
\dontrun{
#' OrgRelLkp(PrimaryRole = "RO197",FromDate="2013-04-01")
}

# return GP Practice to CCG Lookup to include all organisations effective on or after 01-04-2013
\dontrun{
OrgRelLkpGP(PrimaryRole    = "RO177",
          NonPrimaryRole = "RO76",
          RelTypes       = "RE4",
          RelPrimaryRoles = "RO98",
          FromDate       = "2013-04-01")
}

# return GP practice to CCG/PCT Lookup to include all organisations effective on or after 01-04-2012
\dontrun{
OrgRelLkpGP("RO177","RO76","RE4",c("RO98","RO179"),"2012-04-01")
}

}
\seealso{
Other odsR package functions: 
\code{\link{OrgRelLkpNHSTrust}()},
\code{\link{addorgname}()},
\code{\link{getODSfull}()},
\code{\link{getODS}()}
}
\concept{odsR package functions}
