% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{role_nonprimary_lut}
\alias{role_nonprimary_lut}
\title{A lookup of non-primary roles}
\format{A dataframe wiht 55 rows of four variables
\describe{
  \item{code}{A role id}
  \item{role_name}{A role name}
  \item{role_start_date}{Date role came into existence}
  \item{role_end_date}{Date role ceased to exist}
}}
\source{
HSCOrgRefData_DataItemCatalogue.pdf from \url{https://digital.nhs.uk/binaries/content/assets/website-assets/services/ods/xml-organisation-data-products/ancilliary.zip}
}
\usage{
role_nonprimary_lut
}
\description{
A lookup of non-primary roles
}
\keyword{datasets}
