% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{role_primary_lut}
\alias{role_primary_lut}
\title{Look-up table of role id codes and their meanings}
\format{A dataframe of 119 rows and 5 columns:
\describe{
  \item{role_id}{Role Code}
  \item{role_name}{Description of code}
  \item{role_start_date}{Date role started}
  \item{role_end_date}{Date role ended}
  \item{notes}{Notes}
}}
\source{
HSCOrgRefData_DataItemCatalogue.pdf from \url{https://digital.nhs.uk/binaries/content/assets/website-assets/services/ods/xml-organisation-data-products/ancilliary.zip}
}
\usage{
role_primary_lut
}
\description{
A data set containing lookups for role id codes
}
\keyword{datasets}
