% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddOrgName.R
\name{AddOrgName}
\alias{AddOrgName}
\title{AddOrgName}
\usage{
AddOrgName(data, CodeColName, AddedColName)
}
\arguments{
\item{data}{data.frame containing the Organisation Codes to add Organisation Names to}
}
\value{
returns the original data.frame with Organisation Name column inserted immediately after the Organisation Code column
}
\description{
Adds an Organisation Name to a data.frame containing an Organisation Code using the NHS Digitial ODS ORD API.
}
\section{Notes}{
 View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{
# add ParentOrgNM to a GP Practice to CCG Lookup
AddOrgName(lkup,"ParentCD","ParentNM")

}
\seealso{
Other odsR package functions: \code{\link{OrgParentLkp}},
  \code{\link{getODSfull}}, \code{\link{getODS}}
}
