% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getODSfull.R
\name{getODSfull}
\alias{getODSfull}
\title{getODSfull}
\usage{
getODSfull(ODSCode, UseProxy = FALSE)
}
\arguments{
\item{ODSCode}{The organisation to return details for; quoted string; no default}

\item{UseProxy}{Whether to use proxy settings to connect to the directory.spineservices.nhs.uk/ODSAPISuite API.
If TRUE, the function will look for a UseProxy.csv file in the package library extdata directory.
The file must contain a single row of comma separated headers (url, port, username, password, auth)
and a second row specifying the values to be passed to each of these httr::use_proxy() function arguments.
All Argument headers must be present, leave values blank if not required. Quoted string, default FALSE}
}
\value{
returns a list of length 1 containing the full details for the Organisation including:
        Name, Legal and Operational Start and End Dates, Status, Last Change Date, Organisation Record Class,
        Address, Postcode, Contact Information, Roles and Relationships to other Organisations
}
\description{
Extracts full ODS data for a single organisation from the NHS Digital ODS ORD API into a list of length 1.
}
\section{Notes}{
 View the NHS Digital ODS API Implementation Guide at
         \url{https://developer.nhs.uk/library/identifiers/ods-ord-api-implementation-guide/} \cr \cr
         View the NHS Digital ODS API Suite at \url{https://directory.spineservices.nhs.uk/ODSAPISuite}
}

\examples{
# return full Organisation data for RRF12
getODSfull(ODSCode="RRF12")

}
\seealso{
Other odsR package functions: 
\code{\link{OrgRelLkpGP}()},
\code{\link{OrgRelLkpNHSTrust}()},
\code{\link{addorgname}()},
\code{\link{getODS}()}
}
\concept{odsR package functions}
